Remove duplicação de código
