+++ sysarq/src/tests/CreateDocumentSubject.test.js
@@ -3,7 +3,7 @@ import { render, screen, fireEvent, act } from "@testing-library/react";
 import CreateDocumentSubject from "../pages/FieldsRegister/CreateDocumentSubject";
 import { rest } from "msw";
 import { setupServer } from "msw/node";
-import inputChange from "./serverTest";
+import { inputChange } from "./inputTest";
 
 const axiosArchives = `${process.env.REACT_APP_URL_API_ARCHIVES}document-subject/`;
